safari-600.7-branch/LayoutTests/plugins/npruntime/browser-object-identity-expected.txt

Test that plug-in doesn't get a new browser object instance each time

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS embed.getAndForgetRememberedObject().foo is 'bar'
PASS embed.refCount(obj) is 1
PASS embed.refCount(obj) is 1
PASS embed.refCount(obj) is 2
PASS embed.getRememberedObject() is obj
PASS embed.getRememberedObject() is obj
PASS embed.refCount(obj) is 2
PASS embed.getAndForgetRememberedObject() is obj
PASS embed.refCount(obj) is 1
PASS successfullyParsed is true

TEST COMPLETE
